Arbitrage/requirements.txt
requests
openai
python-dotenv
